+++ src-exe/cabal-plan.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE CPP                   #-}
 {-# LANGUAGE DataKinds             #-}
-{-# LANGUAGE DeriveFunctor         #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings     #-}
 {-# LANGUAGE RecordWildCards       #-}
@@ -11,10 +10,9 @@ module Main where
 import           Prelude                     ()
 import           Prelude.Compat
 
-import           Control.Monad.Compat        (ap, forM_, guard, unless, when)
+import           Control.Monad.Compat        (guard, unless, when)
 import           Control.Monad.ST            (runST)
 import           Control.Monad.State.Strict  (StateT, evalStateT, gets, modify')
-import           Control.Monad.Trans.Class   (lift)
 import           Data.Align                  (align)
 import           Data.ByteString             (ByteString)
 import qualified Data.ByteString             as BS
@@ -22,12 +20,11 @@ import qualified Data.ByteString.Lazy        as LBS
 import           Data.Char                   (isAlphaNum)
 import           Data.Foldable               (for_, toList)
 import qualified Data.Graph                  as G
-import qualified Data.List                   as L
 import           Data.Map                    (Map)
 import qualified Data.Map                    as M
 import           Data.Maybe                  (catMaybes, fromMaybe, isJust,
                                               mapMaybe)
-import           Data.Monoid                 (Any (..), Endo (..))
+import           Data.Monoid                 (Any (..))
 import           Data.Semigroup              (Semigroup (..))
 import           Data.Set                    (Set)
 import qualified Data.Set                    as S
@@ -43,29 +40,21 @@ import qualified Data.Vector.Unboxed         as U
 import qualified Data.Vector.Unboxed.Mutable as MU
 import           Data.Version
 import           Flag
-import           GHC.IO.Encoding.Types       (textEncodingName)
 import           Optics.Indexed.Core         (ifor_)
 import           Options.Applicative
 import           ProcessLazyByteString       (readProcessWithExitCode)
-import           System.Console.ANSI
 import           System.Directory            (getCurrentDirectory)
 import           System.Exit                 (ExitCode (..), exitFailure)
-import           System.IO                   (hGetEncoding, hPutStrLn, stderr,
-                                              stdout)
+import           System.IO                   (hPutStrLn, stderr)
 import qualified Text.Parsec                 as P
 import qualified Text.Parsec.String          as P
 import qualified Topograph                   as TG
 
 import           Cabal.Plan
+import           CText
 import           LicenseReport               (generateLicenseReport)
 import           Paths_cabal_plan            (version)
 
-haveUnderlineSupport :: Bool
-#if defined(UNDERLINE_SUPPORT)
-haveUnderlineSupport = True
-#else
-haveUnderlineSupport = False
-#endif
 
 data ShowBuiltin = ShowBuiltin
 data ShowGlobal  = ShowGlobal
@@ -393,8 +382,17 @@ main = do
             <*> switchM DotTredWght "tred-weights" "Adjust edge thickness during transitive reduction"
             <*> many highlightParser
             <*> many (patternOption [ long "root", metavar "PATTERN", help "Graph root(s)", completer $ patternCompleter True ])
-            <*> strOption (mconcat [ short 'o', long "output", metavar "FILE", value "-", showDefault, completer (bashCompleter "file") ])
+            <*> strOption (mconcat [ short 'o', long "output", metavar "FILE", value "-", showDefault, completer (bashCompleter "file"), help "Output file"])
             <*> optional runDot
+            <**> helper
+        , subCommand "dot-png" "Generate dependency PNG image" $ DotCommand
+            <$> planParser
+            <*> pure (toFlag DotTred True)
+            <*> pure (toFlag DotTredWght True)
+            <*> many highlightParser
+            <*> many (patternOption [ long "root", metavar "PATTERN", help "Graph root(s)", completer $ patternCompleter True ])
+            <*> strOption (mconcat [ short 'o', long "output", value "deps.png", metavar "FILE", showDefault, completer (bashCompleter "file"), help "Output file"])
+            <*> pure (Just PNG)
             <**> helper
         , subCommand "topo" "Plan in a topological sort" $ TopoCommand
             <$> planParser
@@ -499,7 +497,7 @@ doFingerprint plan showCabSha = do
 
 doInfo :: UseColors -> UseAscii -> Maybe FilePath -> PlanJson -> IO ()
 doInfo useColors useAscii mProjbase plan = do
-    forM_ mProjbase $ \projbase ->
+    for_ mProjbase $ \projbase ->
         putStrLn ("using '" ++ projbase ++ "' as project root")
     putStrLn ""
     putStrLn "Tree"
@@ -1183,7 +1181,7 @@ doLicenseReport mlicdir pat = do
 
       lst@(_:_:_) -> do
         hPutStrLn stderr "Multiple matching components found:"
-        forM_ lst $ \(pat', uid, cn) -> do
+        for_ lst $ \(pat', uid, cn) -> do
           hPutStrLn stderr ("- " ++ T.unpack pat' ++ "   " ++ show (uid, cn))
         exitFailure
 
@@ -1356,152 +1354,3 @@ graphFromMap m = (g, v2k')
 
     (g, v2k, _) = G.graphFromEdges [ ((), k, S.toList v)
                                    | (k,v) <- M.toList m ]
-
--------------------------------------------------------------------------------
--- Colors and Ascii
--------------------------------------------------------------------------------
-
-data CPiece = CPiece !T [SGR]
-  deriving (Eq, Show)
-
-data T
-    = T !T.Text
-    | Vert -- vertical
-    | Junc -- junction
-    | Corn -- corner
-    | Spac -- space
-    | Rest -- "ellipsis"
-  deriving (Eq, Show)
-
-newtype CText = CText [CPiece]
-  deriving (Eq, Show)
-
-instance IsString CText where
-    fromString s
-        | null s    = mempty
-        | otherwise = CText [CPiece (T (fromString s)) []]
-
-instance Semigroup CText where
-    CText xs <> CText ys = CText (xs <> ys)
-
-instance Monoid CText where
-    mempty  = CText []
-    mappend = (<>)
-
-fromText :: T.Text -> CText
-fromText t = CText [CPiece (T t) []]
-
-fromT :: T -> CText
-fromT t = CText [CPiece t []]
-
-colorifyStr :: Color -> String -> CText
-colorifyStr c t = CText [CPiece (T (T.pack t)) [SetColor Foreground Vivid c]]
-
-colorifyText :: Color -> T.Text -> CText
-colorifyText c t = CText [CPiece (T t) [SetColor Foreground Vivid c]]
-
-recolorify :: Color -> CText -> CText
-recolorify c (CText xs) = CText
-    [ CPiece t (SetColor Foreground Vivid c : sgr)
-    | CPiece t sgr' <- xs
-    , let sgr = filter notSetColor sgr'
-    ]
-  where
-    notSetColor SetColor {} = False
-    notSetColor _           = True
-
--- | We decide to bold, we could do something else to.
-emphasise :: CText -> CText
-emphasise (CText xs) = CText
-    [ CPiece t (SetConsoleIntensity BoldIntensity : sgr)
-    | CPiece t sgr <- xs
-    ]
-
-underline :: CText -> CText
-underline (CText xs) | haveUnderlineSupport  = CText
-    [ CPiece t (SetUnderlining SingleUnderline : sgr)
-    | CPiece t sgr <- xs
-    ]
-underline x = x
-
--- | Colored writer (list is lines)
-newtype CWriter a = CWriter { unCWriter :: Endo [CText] -> (Endo [CText], a) }
-  deriving Functor
-
-class Monad m => MonadCWriter m where
-    putCTextLn :: CText -> m ()
-
-instance MonadCWriter CWriter where
-    putCTextLn t = CWriter $ \l -> (l <> Endo (t :), ())
-
-instance MonadCWriter m => MonadCWriter (StateT s m) where
-    putCTextLn = lift . putCTextLn
-
-instance Applicative CWriter where
-    pure  = return
-    (<*>) = ap
-
-instance Monad CWriter where
-    return x = CWriter $ \ls -> (ls, x)
-
-    m >>= k = CWriter $ \ls0 ->
-        let (ls1, x) = unCWriter m ls0
-        in unCWriter (k x) ls1
-
-data UseColors = ColorsNever | ColorsAuto | ColorsAlways
-  deriving (Eq, Show)
-
-data UseAscii = UseAscii | UseUnicode | UseAsciiAuto
-  deriving (Eq, Show)
-
-runCWriterIO :: UseColors -> UseAscii -> CWriter () -> IO ()
-runCWriterIO ColorsNever  useAscii m = do
-    useAscii' <- shouldUseAscii useAscii
-    runCWriterIONoColors useAscii' m
-runCWriterIO ColorsAlways useAscii m = do
-    useAscii' <- shouldUseAscii useAscii
-    runCWriterIOColors useAscii' m
-runCWriterIO ColorsAuto   useAscii m = do
-    useAscii' <- shouldUseAscii useAscii
-    supports <- hSupportsANSIColor stdout
-    if supports
-    then runCWriterIOColors   useAscii' m
-    else runCWriterIONoColors useAscii' m
-
--- TODO: check environment variables?
-shouldUseAscii :: UseAscii -> IO Bool
-shouldUseAscii UseAscii     = return True
-shouldUseAscii UseUnicode   = return False
-shouldUseAscii UseAsciiAuto = do
-    e <- hGetEncoding stdout
-    return $ not $ fmap (L.isPrefixOf "UTF" . textEncodingName) e == Just True
-
-putT :: Bool -> T -> IO ()
-putT _     (T t) = T.putStr t
--- https://en.wikipedia.org/wiki/Box-drawing_character
-putT False Vert  = T.putStr " \x2502  "
-putT False Junc  = T.putStr " \x251c\x2500 "
-putT False Corn  = T.putStr " \x2514\x2500 "
-putT False Rest  = T.putStr " \x2504\x2504"
--- ascii
-putT True  Vert  = T.putStr " |  "
-putT True  Junc  = T.putStr " +- "
-putT True  Corn  = T.putStr " +- "
-putT True  Rest  = T.putStr " ..."
--- space is just space
-putT _     Spac  = T.putStr "    "
-
-runCWriterIOColors :: Bool -> CWriter () -> IO ()
-runCWriterIOColors useAscii (CWriter f) =
-    forM_ (appEndo (fst (f mempty)) []) $ \(CText l) -> do
-        forM_ l $ \(CPiece t sgr) -> do
-            unless (null sgr) $ setSGR sgr
-            putT useAscii t
-            unless (null sgr) $ setSGR []
-        putChar '\n'
-
-runCWriterIONoColors :: Bool -> CWriter () -> IO ()
-runCWriterIONoColors useAscii (CWriter f) =
-    forM_ (appEndo (fst (f mempty)) []) $ \(CText l) -> do
-        forM_ l $ \(CPiece t _) -> putT useAscii t
-        putChar '\n'